requirements.txt
customtkinter==5.2.1
Pillow==10.0.1
pyperclip==1.8.2
ttkbootstrap==1.10.1